Use QImage for the backing, not QPixmap
one of the oddities of Windows, QPixmap is really slow to change.  QImage is better there.

Difference was ~2-3s on a fast i9-9xxx on Windows 10, vs the 700ms on an i5-6xxx Linux laptop.  The i9 is no doubt the faster system.

Changing to QImage has dropped the execution time to about 1200ms - a bit more reasonable.