Improve Vector inlining and compiler precomputation
By making all the vector classes explicitly constexpr, we let the compiler shortcircuit any constant vector chains in our source by precomputing the value and using the precomputed value as the result.

It may also hint that the function is pure, and can be optimised aggressively on that basis against other constexpr functions - none the less, the improvement in performance is measurable. :)